Move relevant logic into Main component
