+++ src/app/containers/Main.js
@@ -1,9 +1,58 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
+import { PropTypes } from 'prop-types'
+import { connect } from 'react-redux'
+
+import { actions, selectors } from 'pltr/v2'
 
 import App from './App'
 
-const Main = (forceProjectDashboard) => {
+const Main = ({ forceProjectDashboard, busyBooting, setOffline }) => {
+  const [firstTimeBooting, setFirstTimeBooting] = useState(busyBooting)
+
+  // A latch so that we only show initial loading splash once.
+  useEffect(() => {
+    if (!busyBooting && firstTimeBooting) {
+      setFirstTimeBooting(false)
+    }
+  }, [busyBooting])
+
+  useEffect(() => {
+    setOffline(!window.navigator.onLine)
+    const onlineListener = window.addEventListener('online', () => {
+      setOffline(false)
+    })
+    const offlineListener = window.addEventListener('offline', () => {
+      setOffline(true)
+    })
+    return () => {
+      window.removeEventListener('online', onlineListener)
+      window.removeEventListener('offline', offlineListener)
+    }
+  }, [setOffline])
+
+  if (firstTimeBooting) {
+    // TODO: @cameron, @jeana, this is where we can put a more
+    // interesting loading component for users and let them know what
+    // we're loading based on the `applicationState` key in Redux ^_^
+    return (
+      <div id="temporary-inner">
+        <img src="../icons/logo_28_500.png" height="500" />
+      </div>
+    )
+  }
+
   return <App forceProjectDashboard={forceProjectDashboard} />
 }
 
-export default Main
+Main.propTypes = {
+  forceProjectDashboard: PropTypes.bool,
+  busyBooting: PropTypes.bool,
+  setOffline: PropTypes.func.isRequired,
+}
+
+export default connect(
+  (state) => ({
+    busyBooting: selectors.applicationIsBusyAndUninterruptableSelector(state.present),
+  }),
+  { setOffline: actions.project.setOffline }
+)(Main)
